+++ src/app/config/app.config.ts
@@ -2,6 +2,11 @@ export const AppConfig = {
   // Leaflet map
   MAP_CIRCLE_COLOR: '#edd137',
   MAP_CIRCLE_OPACITY: 0.5,
+  MAP_ICONS_URL: {
+    iconRetina: 'assets/icons/marker-icon-2x.png',
+    icon: 'assets/icons/marker-icon.png',
+    shadow: 'assets/icons/marker-shadow.png'
+  },
   MAP_INIT_ACCURACY: 1000,
   MAP_INIT_LATITUDE: 43.604652, // Toulouse
   MAP_INIT_LONGITUDE: 1.444209, // Toulouse